requirements.txt
crewai==1.7.2
folium>=0.15.0
geopy>=2.4.0
langchain_chroma>=0.1.0
langchain_community>=0.2.0
langchain_ollama>=0.1.0
langchain_text_splitters>=0.2.0
litellm<=1.60.0  
pandas==2.3.3
plotly==6.5.1
python-dotenv>=1.1.0
streamlit>=1.30.0
streamlit_folium>=0.15.0
